Pizza rest controller updated

+++ src/main/java/org/lessons/java/spring_la_mia_pizzeria_crud/controllers/PizzaRestController.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.lessons.java.spring_la_mia_pizzeria_crud.model.Pizza;
 import org.lessons.java.spring_la_mia_pizzeria_crud.service.PizzaService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,6 +15,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping("/api/pizzas")
@@ -28,23 +32,33 @@ public class PizzaRestController {
     }
     // SHOW--------------------------------
     @GetMapping("/{id}")
-    public Pizza show(@PathVariable Integer id) {
-        return pizzaService.getById(id);
+    public ResponseEntity<Pizza> show(@Valid @PathVariable Integer id) {
+        if (pizzaService.findById(id).isEmpty()) {
+            return new ResponseEntity<Pizza>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<Pizza>(pizzaService.getById(id), HttpStatus.OK);
     }
     // STORE------------------------------
     @PostMapping
-    public Pizza store(@RequestBody Pizza pizza) {
-        return pizzaService.create(pizza);
+    public ResponseEntity<Pizza> store(@Valid @RequestBody Pizza pizza) {
+        return new ResponseEntity<Pizza>(pizzaService.create(pizza), HttpStatus.OK);
     }
     // UPDATE-----------------------------
     @PutMapping("/{id}")
-    public Pizza update(@RequestBody Pizza pizza, @PathVariable Integer id) {
+    public ResponseEntity<Pizza> update(@Valid @RequestBody Pizza pizza, @Valid @PathVariable Integer id) {
+        if (pizzaService.findById(id).isEmpty()) {
+            return new ResponseEntity<Pizza>(HttpStatus.NOT_FOUND);
+        }
         pizza.setId(id);
-        return pizzaService.update(pizza);
+        return new ResponseEntity<Pizza>(pizzaService.update(pizza), HttpStatus.OK);
     }
     // DELETE----------------------------
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Integer id) {
+    public ResponseEntity<Pizza> delete(@Valid @PathVariable Integer id) {
+        if (pizzaService.findById(id).isEmpty()) {
+            return new ResponseEntity<Pizza>(HttpStatus.NOT_FOUND);
+        }
         pizzaService.deleteById(id);
+        return new ResponseEntity<Pizza>(HttpStatus.OK);
     }
 }
